backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic==2.9.2
pytesseract==0.3.13
Pillow==10.4.0
requests==2.32.3
langchain==0.3.7
langchain-google-genai==2.0.1
langchain-community==0.3.7
pydantic-core==2.23.4
